Set up site for publication
Adding `_config.yml` and `index.md` was to confirm that the site compiles correctly. The `.gitignore` update ensures that the compiled `_site/` directory doesn't get tracked in version control since it would cause too many potential conflicts when multiple editors are working on it.

Instead of keeping the compiled `_site/` in version control our deployment pipeline will have to include a jekyll build process that creates that directory on the host.